feat: fetching user listing

+++ server/controller/user.controller..js
@@ -31,9 +31,9 @@ export const updateUser = async (req, res, next) => {
 };
 
 export const deleteUser = async (req, res, next) => {
-    // if (req.user.id !== req.params.id) {
-    //     return next(errorHandler(401, 'You can only delete your account'));
-    // }
+    if (req.user.id !== req.params.id) {
+        return next(errorHandler(401, 'You can only delete your account'));
+    }
     
     try {
      await User.findByIdAndDelete(req.params.id);
